% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{add_cohort_start_date}
\alias{add_cohort_start_date}
\title{Add cohort start date}
\usage{
add_cohort_start_date(data)
}
\arguments{
\item{data}{A data frame containing a COHORT_startjaar column}
}
\value{
The input data frame with an additional COHORT_start_datum column
containing dates set to August 1st of the start year
}
\description{
Adds a start date column to cohort data based on the start year
}
