% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{add_cohort_start_year}
\alias{add_cohort_start_year}
\title{Add Cohort Start Year}
\usage{
add_cohort_start_year(data)
}
\arguments{
\item{data}{A data frame containing a COHORT_naam column}
}
\value{
The input data frame with an additional COHORT_startjaar column containing
numeric start years
}
\description{
Extract and add the start year from cohort names
}
