% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{add_filled_booleans}
\alias{add_filled_booleans}
\title{Add Filled Boolean Columns}
\usage{
add_filled_booleans(data, fill = FALSE, text = "opgevuld")
}
\arguments{
\item{data}{A data frame containing logical columns with "\emph{is}" in their names}

\item{fill}{Optional. A logical value to replace NA values with (default: FALSE)}

\item{text}{Optional. A string to append to the new column names (default: "opgevuld")}
}
\value{
A data frame with additional boolean columns. New columns are named as
original_name_text where NA values are replaced with the fill value.
}
\description{
Create new boolean columns with NA values replaced by a specified value
}
