% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{add_grouped_enrollment_dates}
\alias{add_grouped_enrollment_dates}
\title{Add First and Last Enrollment Dates}
\usage{
add_grouped_enrollment_dates(enrollments)
}
\arguments{
\item{enrollments}{A data frame containing enrollment data with columns
VERBINTENIS_volgnummer, VERBINTENIS_begindatum, and VERBINTENIS_einddatum}
}
\value{
The input data frame with two additional columns:
\itemize{
\item VERBINTENIS_begindatum_eerst: The earliest start date per VERBINTENIS_volgnummer
\item VERBINTENIS_einddatum_laatst: The latest end date per VERBINTENIS_volgnummer
}
}
\description{
Calculate the first and last enrollment dates for each enrollment sequence
}
