% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{add_school_year}
\alias{add_school_year}
\title{Add School Year Column}
\usage{
add_school_year(data)
}
\arguments{
\item{data}{A data frame containing the column MEDEWERKER_contract_fte_peildatum}
}
\value{
The input data frame with an additional SCHOOLJAAR_naam column in the format "YYYY/YYYY+1"
}
\description{
Adds a school year column based on a reference date
}
