% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{calculate_application_duration}
\alias{calculate_application_duration}
\title{Calculate the Duration of Application Process}
\usage{
calculate_application_duration(data)
}
\arguments{
\item{data}{A data frame containing application data with columns
AANMELDING_begin_datum and AANMELDING_laatst_gewijzigd_datum}
}
\value{
A data frame with an additional column AANMELDING_in_proces
containing the rounded number of days between start and last modification
}
\description{
Takes a data frame containing application data and calculates the number of days
between the application start date and last modification date.
}
\details{
The function calculates the difference in days between AANMELDING_laatst_gewijzigd_datum
and AANMELDING_begin_datum, rounds the result, and adds it as a new column.
}
\examples{
\dontrun{
enrollments_application <- calculate_application_duration(raw_enrollments_application)
}
}
