% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{calculate_application_to_start}
\alias{calculate_application_to_start}
\title{Calculate Days Between Application and Start Date}
\usage{
calculate_application_to_start(data)
}
\arguments{
\item{data}{A data frame containing columns:
\itemize{
\item AANMELDING_begin_datum: Application start date
\item AANMELDING_laatst_gewijzigd_datum: Last application modification date
\item VERBINTENIS_begindatum: Enrollment start date
}}
}
\value{
The input data frame with two additional columns:
\itemize{
\item AANMELDING_begin_dagen_tot_start: Days between application and start
\item AANMELDING_afgerond_dagen_tot_start: Days between last modification and start
}
Values greater than 365 days or less than 0 days are set to NA.
}
\description{
Calculates the number of days between application dates and enrollment start dates
}
