% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{calculate_exam_plan_to_specific_dates}
\alias{calculate_exam_plan_to_specific_dates}
\title{Calculate Exam Plan Date Comparisons}
\usage{
calculate_exam_plan_to_specific_dates(data)
}
\arguments{
\item{data}{A dataframe containing exam plan dates and COHORT_startj}
}
\description{
Adds columns comparing exam plan dates to April 1st of the cohort's start year
}
