% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{calculate_is_havo_vwo_dropout}
\alias{calculate_is_havo_vwo_dropout}
\title{Identify Failed HAVO/VWO Students}
\usage{
calculate_is_havo_vwo_dropout(enrollments)
}
\arguments{
\item{enrollments}{A data frame containing prior education columns}
}
\value{
A data frame with an additional column DEELNEMER_havo_vwo_gezakt indicating
if the student failed HAVO/VWO
}
\description{
Determine if students failed HAVO/VWO based on prior education data
}
