% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{correlate_teams}
\alias{correlate_teams}
\title{Calculate Correlations Between Team Variables}
\usage{
correlate_teams(teams)
}
\arguments{
\item{teams}{A data frame containing team-related numeric variables}
}
\value{
A correlation matrix of numeric variables, with NA values replaced by 0
and rearranged for better visualization.
}
\description{
Calculate correlations between numeric variables in team data
}
