% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{create_cv_folds}
\alias{create_cv_folds}
\title{Create Cross-Validation Folds for Model Training}
\usage{
create_cv_folds(data, outcome_var = "DEELNEMER_BC_uitval", n_folds = 5)
}
\arguments{
\item{data}{Training data}

\item{outcome_var}{Name of the outcome variable}

\item{n_folds}{Number of folds for cross-validation}
}
\value{
A vfold_cv object for use in tuning
}
\description{
Create stratified cross-validation folds for model tuning.
}
