% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{create_model_recipe}
\alias{create_model_recipe}
\title{Create Model Recipe for Predictive Modeling}
\usage{
create_model_recipe(data, outcome_var = "DEELNEMER_BC_uitval")
}
\arguments{
\item{data}{Training data}

\item{outcome_var}{Name of the outcome variable}
}
\value{
A recipe object for preprocessing data
}
\description{
Create a recipe for data preprocessing in the modeling pipeline.
}
