% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{create_model_workflow}
\alias{create_model_workflow}
\title{Create Model Workflow}
\usage{
create_model_workflow(model_spec, model_recipe)
}
\arguments{
\item{model_spec}{Model specification from create_model_spec()}

\item{model_recipe}{Recipe from create_model_recipe()}
}
\value{
A workflow object
}
\description{
Create a workflow combining model specification and recipe.
}
