% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{create_special_needs_boolean}
\alias{create_special_needs_boolean}
\title{Create Boolean for Special Educational Needs}
\usage{
create_special_needs_boolean(data)
}
\arguments{
\item{data}{A data frame containing a VERBINTENIS_passend_onderwijs_kenmerk column}
}
\value{
The input data frame with an additional boolean column VERBINTENIS_is_passend_onderwijs
indicating if a student has special educational needs.
}
\description{
Creates a boolean indicator for students with special educational needs
}
