% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{create_student_staff_ratio}
\alias{create_student_staff_ratio}
\title{Calculate Student-Staff Ratio}
\usage{
create_student_staff_ratio(teams)
}
\arguments{
\item{teams}{A data frame containing team data with columns TEAM_studenten_aantal and MEDEWERKER_contract_fte_totaal.}
}
\value{
The input data frame with an additional column TEAM_student_staf_ratio.
}
\description{
Calculate the ratio between students and staff FTE per team
}
