% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prepare.R
\name{encode_model_factors}
\alias{encode_model_factors}
\title{Encode Factors for Modeling}
\usage{
encode_model_factors(data, outcome_var = "DEELNEMER_BC_uitval")
}
\arguments{
\item{data}{A data frame to process}

\item{outcome_var}{The name of the outcome variable}
}
\value{
A data frame with properly encoded factors and handled missing values
}
\description{
Encode categorical variables as factors and handle missing values.
}
