% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{filter_employees_to_education_teams_only}
\alias{filter_employees_to_education_teams_only}
\title{Filter Employees to Education Teams Only}
\usage{
filter_employees_to_education_teams_only(employees)
}
\arguments{
\item{employees}{A dataframe of employees containing TEAM_kostenplaats_code}
}
\value{
A filtered dataframe containing only employees in education teams
}
\description{
Filter the employees dataset to only include those in education teams
}
