% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{finalize_model_workflow}
\alias{finalize_model_workflow}
\title{Finalize Model Workflow with Best Parameters}
\usage{
finalize_model_workflow(
  model_workflow,
  best_params,
  model_type = "random_forest"
)
}
\arguments{
\item{model_workflow}{Workflow from create_model_workflow()}

\item{best_params}{Best parameters from select_best_params()}

\item{model_type}{Type of model}
}
\value{
A finalized workflow
}
\description{
Finalize a workflow with the best hyperparameters.
}
