% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{fix_inf_values}
\alias{fix_inf_values}
\title{Replace infinite values with NA}
\usage{
fix_inf_values(data)
}
\arguments{
\item{data}{A data frame or tibble.}
}
\value{
A data frame with the same structure as the input, but with infinite values
replaced by NA in all numeric columns.
}
\description{
Replace infinite values in numeric columns with NA
}
