% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{format_school_year_name}
\alias{format_school_year_name}
\title{Format School Year Names}
\usage{
format_school_year_name(teams_results)
}
\arguments{
\item{teams_results}{A data frame containing a column SCHOOLJAAR_naam_met_streep}
}
\value{
A data frame with formatted school year names in new column SCHOOLJAAR_naam
}
\description{
Convert school year names from a dash format to a slash format
}
