% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prepare.R
\name{get_tbl_summary}
\alias{get_tbl_summary}
\title{Create a Summary Table for Student Retention Analysis}
\usage{
get_tbl_summary(df)
}
\arguments{
\item{df}{A data frame containing student data with a 'Retentie' column and various
student characteristics to be summarized}
}
\value{
A object containing the formatted summary statistics, including means
and standard deviations for continuous variables, counts and percentages for
categorical variables, with p-values and Bonferroni-adjusted q-values
}
\description{
Creates a formatted summary table comparing student characteristics against retention status
}
