% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_crebo_to_organisation_mapping}
\alias{ingest_crebo_to_organisation_mapping}
\title{Ingest CREBO to Organization Mapping Data}
\usage{
ingest_crebo_to_organisation_mapping(
  ...,
  filename = NULL,
  path = NULL,
  config_key = "crebo_to_organisation_mapping",
  config_data_path = "data_raw_dir"
)
}
\arguments{
\item{...}{Additional arguments passed to load_data().}

\item{filename}{Optional. A single string for the input file name.}

\item{path}{Optional. A single string for the file path.}

\item{config_key}{Optional. A single string for the configuration key (default: "crebo_to_organisation_mapping").}

\item{config_data_path}{Optional. A single string for the raw data path in config (default: "data_raw_dir").}
}
\value{
A tibble with cleaned column names, all prefixed with "OPLEIDING_".
}
\description{
Load and process CREBO (Dutch education code) to organization mapping data from a file
}
