% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_employee_answers_satisfaction_2020}
\alias{ingest_employee_answers_satisfaction_2020}
\title{Ingest Employee Satisfaction Survey Answers for 2020}
\usage{
ingest_employee_answers_satisfaction_2020(
  ...,
  filename = NULL,
  path = NULL,
  config_key = "employees_satisfaction_2020",
  config_data_path = "data_raw_dir"
)
}
\arguments{
\item{...}{Additional arguments passed to the helper function.}

\item{filename}{Optional. A single string for the filename.}

\item{path}{Optional. A single string for the file path.}

\item{config_key}{Optional. A single string specifying the configuration key (default: "employees_satisfaction_2020").}

\item{config_data_path}{Optional. A single string specifying the config path for raw data (default: "data_raw_dir").}
}
\value{
A tibble containing processed employee satisfaction survey responses.
}
\description{
Import and process employee satisfaction survey responses from the 2020 survey
}
