% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_employee_answers_satisfaction_2022}
\alias{ingest_employee_answers_satisfaction_2022}
\title{Ingest Employee Satisfaction Survey Answers 2022}
\usage{
ingest_employee_answers_satisfaction_2022(
  ...,
  filename = NULL,
  path = NULL,
  config_key = "employees_satisfaction_2022",
  config_data_path = "data_raw_dir"
)
}
\arguments{
\item{...}{Additional arguments passed to the underlying helper function.}

\item{filename}{Optional. A string specifying the name of the input file.}

\item{path}{Optional. A string specifying the path to the input file.}

\item{config_key}{Optional. A string specifying the configuration key (default: "employees_satisfaction_2022").}

\item{config_data_path}{Optional. A string specifying the config path for raw data (default: "data_raw_dir").}
}
\value{
A cleaned data frame containing employee satisfaction survey responses.
}
\description{
Read and process employee satisfaction survey answers from 2022
}
