% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_employee_answers_satisfaction_codebook_2024}
\alias{ingest_employee_answers_satisfaction_codebook_2024}
\title{Ingest Employee Satisfaction Survey Codebook Data for 2024}
\usage{
ingest_employee_answers_satisfaction_codebook_2024(
  ...,
  filename = NULL,
  path = NULL,
  config_key = "employees_satisfaction_codebook_2024",
  config_data_path = "data_raw_dir"
)
}
\arguments{
\item{...}{Additional arguments passed to the underlying data loading function.}

\item{filename}{Optional. A string specifying the name of the file to read.}

\item{path}{Optional. A string specifying the path to the file.}

\item{config_key}{Optional. A string specifying the configuration key to use (default: "employees_satisfaction_codebook_2024").}

\item{config_data_path}{Optional. A string specifying the config path for raw data (default: "data_raw_dir").}
}
\value{
A tibble containing processed employee satisfaction codebook data.
}
\description{
Reads and processes employee satisfaction survey codebook data from a file
}
