% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_employee_answers_satisfaction_codebook_helper}
\alias{ingest_employee_answers_satisfaction_codebook_helper}
\title{Ingest Employee Satisfaction Survey Codebook}
\usage{
ingest_employee_answers_satisfaction_codebook_helper(
  config_key,
  ...,
  filename = NULL,
  path = NULL,
  config_data_path = "data_raw_dir"
)
}
\arguments{
\item{config_key}{A string indicating the configuration key to use.}

\item{...}{Additional arguments passed to load_data.}

\item{filename}{Optional. A string specifying the input file name.}

\item{path}{Optional. A string specifying the input file path.}

\item{config_data_path}{Optional. A string specifying the configuration data path.}
}
\value{
Returns the processed data frame with QuestionId as character and other columns
guessed. The data is also saved using save_ingested().
}
\description{
Reads and processes employee satisfaction survey codebook data with specific column type handling
}
