% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_employees_job_type_2019}
\alias{ingest_employees_job_type_2019}
\title{Ingest Employee Job Type Data from 2019}
\usage{
ingest_employees_job_type_2019(
  ...,
  filename = NULL,
  path = NULL,
  config_key = "employees_job_type_2019",
  config_data_path = "data_raw_dir"
)
}
\arguments{
\item{...}{Additional arguments passed to readr::read_delim.}

\item{filename}{Optional. A string specifying the name of the CSV file to read.}

\item{path}{Optional. A string specifying the path to the CSV file.}

\item{config_key}{Optional. A string specifying the configuration key to use.}

\item{config_data_path}{Optional. A string specifying the config path for raw data.}
}
\value{
A tibble containing processed employee job type data.
}
\description{
Reads and processes employee job type data from 2019 from a CSV file
}
