% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_employees_job_type_2021}
\alias{ingest_employees_job_type_2021}
\title{Ingest Employee Job Type Data for 2021}
\usage{
ingest_employees_job_type_2021(
  ...,
  filename = NULL,
  path = NULL,
  config_key = "employees_job_type_2021",
  config_data_path = "data_raw_dir"
)
}
\arguments{
\item{...}{Additional arguments passed to the underlying data reading function.}

\item{filename}{Optional. A single string specifying the CSV file name.}

\item{path}{Optional. A single string specifying the file path.}

\item{config_key}{Optional. A single string specifying the configuration key to use.}

\item{config_data_path}{Optional. A single string specifying the config path for raw data.}
}
\value{
A tibble containing processed employee job type data.
}
\description{
Read and process employee job type data from a CSV file for the year 2021
}
