% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_employees_job_type_2022}
\alias{ingest_employees_job_type_2022}
\title{Ingest Employee Job Type Data for 2022}
\usage{
ingest_employees_job_type_2022(
  ...,
  filename = NULL,
  path = NULL,
  config_key = "employees_job_type_2022",
  config_data_path = "data_raw_dir"
)
}
\arguments{
\item{...}{Additional arguments passed to underlying read functions.}

\item{filename}{Optional. A string specifying the name of the CSV file.}

\item{path}{Optional. A string specifying the path to the CSV file.}

\item{config_key}{Optional. A string specifying the configuration key (defaults to "employees_job_type_2022").}

\item{config_data_path}{Optional. A string specifying the config path for raw data (defaults to "data_raw_dir").}
}
\value{
A tibble containing processed employee job type data.
}
\description{
Read and process employee job type data from a CSV file for the year 2022
}
