% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_employees_job_type_2023}
\alias{ingest_employees_job_type_2023}
\title{Ingest Employee Job Type Data for 2023}
\usage{
ingest_employees_job_type_2023(
  ...,
  filename = NULL,
  path = NULL,
  config_key = "employees_job_type_2023",
  config_data_path = "data_raw_dir"
)
}
\arguments{
\item{...}{Additional arguments passed to the underlying data reading function.}

\item{filename}{Optional. A string specifying the input file name.}

\item{path}{Optional. A string specifying the path to the input file.}

\item{config_key}{Optional. A string specifying the configuration key (default: "employees_job_type_2023").}

\item{config_data_path}{Optional. A string specifying the configuration data path (default: "data_raw_dir").}
}
\value{
A tibble containing processed employee job type data.
}
\description{
Read and process employee job type data from a data source for the year 2023
}
