% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_enrollments_flex}
\alias{ingest_enrollments_flex}
\title{Ingest Flex Status Information}
\usage{
ingest_enrollments_flex(
  ...,
  filename = NULL,
  path = NULL,
  config_key = "enrollments_flex",
  config_data_path = "data_raw_dir"
)
}
\arguments{
\item{...}{Additional arguments passed to readr::read_delim}

\item{filename}{Character string specifying the name of the CSV file to read}
}
\value{
A tibble containing processed flex status data with columns:
\itemize{
\item VERBINTENIS_ID: Unique connection identifier
\item VERBINTENIS_Is_flex: Boolean indicating flex status
}
}
\description{
Reads and processes flex status data from a CSV file.
Expects CSV files with semicolon (;) as separator.
}
