% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_enrollment_to_sk_mapping}
\alias{ingest_enrollment_to_sk_mapping}
\title{Ingest enrollment to SK identifier mapping data}
\usage{
ingest_enrollment_to_sk_mapping(
  ...,
  filename = NULL,
  path = NULL,
  config_key = "enrollment_to_sk_mapping",
  config_data_path = "data_raw_dir"
)
}
\arguments{
\item{...}{Additional arguments passed to readr::read_delim.}

\item{filename}{Optional. A single string specifying the CSV file name.}

\item{path}{Optional. A single string specifying the file path.}

\item{config_key}{Optional. A single string specifying the configuration key (default: "enrollment_to_sk_mapping").}

\item{config_data_path}{Optional. A single string specifying the config path for raw data (default: "data_raw_dir").}
}
\value{
A tibble containing processed mapping data between enrollment and SK identifiers.
}
\description{
Read and process mapping data between enrollment identifiers and SK identifiers
}
