% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_enrollments_special_needs}
\alias{ingest_enrollments_special_needs}
\title{Ingest Special Educational Needs Data}
\usage{
ingest_enrollments_special_needs(
  ...,
  filename = NULL,
  path = NULL,
  config_key = "enrollments_special_needs",
  config_data_path = "data_raw_dir"
)
}
\arguments{
\item{...}{Additional arguments passed to the data loading function.}

\item{filename}{Optional. A string specifying the name of the CSV file to read.}

\item{path}{Optional. A string specifying the path to the CSV file.}

\item{config_key}{Optional. A string specifying the configuration key (defaults to "enrollments_special_needs").}

\item{config_data_path}{Optional. A string specifying the config path for raw data (defaults to "data_raw_dir").}
}
\value{
A tibble containing:
\itemize{
\item VERBINTENIS_ID: Enrollment identifier
\item VERBINTENIS_passend_onderwijs_kenmerk: Special educational needs characteristic
}
}
\description{
Reads and processes data about students with special educational needs from a CSV file
}
