% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_programmes_basic}
\alias{ingest_programmes_basic}
\title{Ingest Education Programme Information}
\usage{
ingest_programmes_basic(
  ...,
  filename = NULL,
  path = NULL,
  config_key = "programmes_basic",
  config_data_path = "data_raw_dir"
)
}
\arguments{
\item{...}{Additional arguments passed to readr::read_delim}

\item{filename}{Character string specifying the name of the CSV file to read}
}
\value{
A tibble containing processed programme data with columns prefixed with OPLEIDING_
}
\description{
Reads and processes education programme data from a CSV file.
Expects CSV files with semicolon (;) as separator.
}
