% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_student_postcode_ses}
\alias{ingest_student_postcode_ses}
\title{Ingest Student Postal Code SES Data}
\usage{
ingest_student_postcode_ses(
  ...,
  filename = NULL,
  path = NULL,
  config_key = "student_postcode_ses",
  config_data_path = "data_raw_dir"
)
}
\arguments{
\item{...}{Additional arguments passed to the data loading function.}

\item{filename}{Optional. A string specifying the name of the CSV file to read.}

\item{path}{Optional. A string specifying the path to the CSV file.}

\item{config_key}{Optional. A string specifying the configuration key (defaults to "student_postcode_ses").}

\item{config_data_path}{Optional. A string specifying the config path for raw data (defaults to "data_raw_dir").}
}
\value{
A tibble containing:
\itemize{
\item DEELNEMER_postcode4: 4-digit postal code
\item DEELNEMER_postcode4_ses_score: The average socioeconomic status score for the postal code
\item DEELNEMER_postcode4_ses_spreiding: The spread (variation) of SES scores within the postal code
}
}
\description{
Reads and processes data about postal codes and socioeconomic status (SES) from a CSV file.
This data includes SES scores for each postal code area based on CBS (Central Bureau of Statistics) data.
}
