% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_students_satisfaction_2021}
\alias{ingest_students_satisfaction_2021}
\title{Ingest Student Satisfaction Data for 2021}
\usage{
ingest_students_satisfaction_2021(
  ...,
  filename = NULL,
  path = NULL,
  config_key = "students_satisfaction_2021",
  config_data_path = "data_raw_dir"
)
}
\arguments{
\item{...}{Additional arguments passed to the underlying data reading function}

\item{filename}{Optional. A string specifying the input file name}

\item{path}{Optional. A string specifying the file path}

\item{config_key}{Optional. A string specifying the configuration key (defaults to "students_satisfaction_2021")}

\item{config_data_path}{Optional. A string specifying the configuration data path (defaults to "data_raw_dir")}
}
\value{
A cleaned data frame containing student satisfaction data
}
\description{
Read and process student satisfaction survey data from 2021
}
