% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_students_satisfaction_2023}
\alias{ingest_students_satisfaction_2023}
\title{Ingest Student Satisfaction Data for 2023}
\usage{
ingest_students_satisfaction_2023(
  ...,
  filename = NULL,
  path = NULL,
  config_key = "students_satisfaction_2023",
  config_data_path = "data_raw_dir"
)
}
\arguments{
\item{...}{Additional arguments passed to readr::read_delim}

\item{filename}{Optional. A string specifying the input file name.}

\item{path}{Optional. A string specifying the path to the input file.}

\item{config_key}{Optional. A string specifying the configuration key (defaults to "students_satisfaction_2023").}

\item{config_data_path}{Optional. A string specifying the config path for raw data (defaults to "data_raw_dir").}
}
\value{
A tibble containing processed student satisfaction data, with CREBO as integer
and studentenhuisvesting_05 as integer columns, and other columns guessed.
}
\description{
Read and process student satisfaction survey data for the year 2023
}
