% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interpret.R
\name{interpret_group_models}
\alias{interpret_group_models}
\title{Interpret Results for Multiple Groups}
\usage{
interpret_group_models(
  model_results_list,
  n_vars = 10,
  save = TRUE,
  use_friendly_names = TRUE
)
}
\arguments{
\item{model_results_list}{A list of model results from run_model() or run_group_models()}

\item{n_vars}{Number of top variables to extract (default: 10)}

\item{save}{Logical indicating whether to save the interpretations (default: TRUE)}

\item{use_friendly_names}{Logical indicating whether to use user-friendly variable names (default: TRUE)}
}
\value{
A list of interpretation results for each group
}
\description{
Run interpretation for multiple model results.
}
