% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{parse_enrollment_dates}
\alias{parse_enrollment_dates}
\title{Parse Enrollment Dates}
\usage{
parse_enrollment_dates(enrollments)
}
\arguments{
\item{enrollments}{A data frame containing enrollment dates in columns
'VERBINTENIS_begindatum' and 'VERBINTENIS_einddatum' in "dd-mm-yyyy" format.}
}
\value{
A data frame with dates converted to Date class.
}
\description{
Convert enrollment start and end dates from character to Date format
}
