% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{parse_enrollment_level}
\alias{parse_enrollment_level}
\title{Parse Enrollment Level}
\usage{
parse_enrollment_level(data)
}
\arguments{
\item{data}{A data frame containing a column \code{VERBINTENIS_niveau_omschrijving}}
}
\value{
A data frame with an additional column \code{VERBINTENIS_niveau} containing the numeric level
}
\description{
Extract numeric level from enrollment level description
}
