% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{parse_result_pct}
\alias{parse_result_pct}
\title{Parse Team Result Percentages}
\usage{
parse_result_pct(teams_results)
}
\arguments{
\item{teams_results}{A data frame containing team results with a column
TEAM_startersresultaat_1_jaars_omschrijving}
}
\value{
A data frame with an additional numeric column TEAM_startersresultaat_1_jaars
containing percentages multiplied by 100
}
\description{
Convert text-based result descriptions to numeric percentages
}
