% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{pivot_answers_to_students}
\alias{pivot_answers_to_students}
\title{Convert Answer Data to Student-Level Format}
\usage{
pivot_answers_to_students(student_answers_satisfaction)
}
\arguments{
\item{student_answers_satisfaction}{A dataframe containing student satisfaction survey responses
in long format.}
}
\value{
A dataframe in wide format where each row represents an student's responses,
with questions as columns and scores as values.
}
\description{
Restructures student satisfaction survey data from long to wide format
}
