% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_with_friendly_names}
\alias{replace_with_friendly_names}
\title{Replace technical variable names with user-friendly names}
\usage{
replace_with_friendly_names(df)
}
\arguments{
\item{df}{Data frame with technical column names}
}
\value{
Data frame with user-friendly column names
}
\description{
Replace technical variable names with user-friendly names
}
