% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{save_model_results}
\alias{save_model_results}
\title{Save Model Results}
\usage{
save_model_results(
  model_results,
  program_filter = NULL,
  level_filter = NULL,
  custom_path = NULL
)
}
\arguments{
\item{model_results}{Model results from run_model()}

\item{program_filter}{The program filter used (or NULL)}

\item{level_filter}{The level filter used (or NULL)}

\item{custom_path}{Optional custom path to save the results}
}
\value{
Invisibly returns the paths where results were saved
}
\description{
Save model results to the appropriate location.
}
