% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{select_cols_for_correlation}
\alias{select_cols_for_correlation}
\title{Select columns for correlation analysis}
\usage{
select_cols_for_correlation(df, na_thresh = 0.8)
}
\arguments{
\item{df}{A data frame.}

\item{na_thresh}{Optional. A numeric value between 0 and 1 indicating the threshold for NA values
(default: 0.8).}
}
\value{
A data frame with columns removed that either have more NAs than the threshold
or only one unique value.
}
\description{
Remove columns with too many missing values or no variation
}
