% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{set_user_friendly_names}
\alias{set_user_friendly_names}
\title{Make Column Names User-Friendly}
\usage{
set_user_friendly_names(teams_correlations)
}
\arguments{
\item{teams_correlations}{A data frame containing column names with 'VERBINTENIS_' and 'TEAM_' prefixes}
}
\value{
A data frame with cleaned column names, removing 'VERBINTENIS_' and 'TEAM_' prefixes
and replacing underscores with spaces
}
\description{
Removes technical prefixes and underscores from column names to make them more readable
}
