% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prepare.R
\name{split_model_data}
\alias{split_model_data}
\title{Split Data into Training and Test Sets}
\usage{
split_model_data(data, cohort_var = "COHORT_startjaar", test_cohort = 2023)
}
\arguments{
\item{data}{A data frame to split}

\item{cohort_var}{The name of the cohort year variable}

\item{test_cohort}{The cohort year to use as test data}
}
\value{
A list containing the training and test datasets
}
\description{
Split data into training and test sets based on cohort year.
}
