% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{summarise_employee_absence_to_years}
\alias{summarise_employee_absence_to_years}
\title{Summarise Employee Absences by School Year}
\usage{
summarise_employee_absence_to_years(employee_absences)
}
\arguments{
\item{employee_absences}{A data frame containing employee absence records}
}
\value{
A data frame containing yearly summarized absence data for employees
}
\description{
Transforms employee absence data into yearly summaries based on school years
}
