% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{summarise_satisfaction_to_groups}
\alias{summarise_satisfaction_to_groups}
\title{Summarize Employee Satisfaction by Groups}
\usage{
summarise_satisfaction_to_groups(employee_answers_satisfaction)
}
\arguments{
\item{employee_answers_satisfaction}{A data frame containing employee satisfaction data}
}
\value{
A data frame grouped by SCHOOLJAAR_startjaar, Organisatie, Characteristic 1,
Characteristic 2, and group_number, with count information removed.
}
\description{
Groups employee satisfaction data by school year, organization, and characteristics
}
