% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_advices_to_enrollments}
\alias{transform_advices_to_enrollments}
\title{Transform Advice Data to Enrollment Format}
\usage{
transform_advices_to_enrollments(study_advices)
}
\arguments{
\item{study_advices}{A data frame containing advice records with the following required columns:
\itemize{
\item VERBINTENIS_ID: Enrollment identifier
\item VERBINTENIS_advies_voorlopig: Binary indicator for preliminary advice (1 = yes)
\item VERBINTENIS_advies_definitief: Binary indicator for definitive advice (1 = yes)
\item VERBINTENIS_advies_aanmaakdatum: Creation date of the advice
}}
}
\value{
A data frame with the following columns:
\itemize{
\item VERBINTENIS_ID: Original enrollment identifier
\item verbintenis_advies_voorlopig_eerste_datum: Date of first preliminary advice
\item verbintenis_advies_definitief_eerste_datum: Date of first definitive advice
}
}
\description{
Transforms advice data into an enrollment format by processing preliminary and
definitive advice records. The function groups data by enrollment ID and advice type,
then pivots the data to create separate columns for different advice types with their
respective creation dates.
}
\details{
The function performs the following transformations:
\enumerate{
\item Determines advice type (preliminary or definitive) based on binary indicators
\item Groups data by enrollment ID and advice type
\item Takes the earliest date for each advice type
\item Pivots the data to create separate columns for each advice type
}
}
\examples{
\dontrun{
advice_data <- data.frame(
  VERBINTENIS_ID = c(1, 1, 2),
  VERBINTENIS_advies_voorlopig = c(1, 0, 1),
  VERBINTENIS_advies_definitief = c(0, 1, 0),
  VERBINTENIS_advies_aanmaakdatum = as.Date(c("2024-01-01", "2024-02-01", "2024-01-15"))
)

result <- transform_advices_to_enrollments(advice_data)
}

}
