% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_cat_val_to_pct_columns}
\alias{transform_cat_val_to_pct_columns}
\title{Transform Categorical Values to Percentage Columns}
\usage{
transform_cat_val_to_pct_columns(data, grouping_vars, max_n_values, min_pct)
}
\arguments{
\item{data}{A data frame containing categorical variables.}

\item{grouping_vars}{Optional. A character vector of column names to group by. Defaults to \code{c("TEAM_naam", "COHORT_naam")}.}

\item{max_n_values}{Optional. Maximum number of unique values for a column to be considered categorical. Defaults to 6.}

\item{min_pct}{Optional. Minimum percentage threshold for including categories. Defaults to 0.10.}
}
\value{
A data frame with categorical variables transformed into percentage columns.
Will error if no variables meet the selection criteria.
}
\description{
Convert categorical variables into percentage columns based on grouping variables
}
