% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_employees_cat_val_to_pct_columns}
\alias{transform_employees_cat_val_to_pct_columns}
\title{Transform Employee Categories to Percentage Columns}
\usage{
transform_employees_cat_val_to_pct_columns(employees)
}
\arguments{
\item{employees}{A data frame containing employee data with columns for team cost center code and school year.}
}
\value{
A data frame with categorical variables transformed into percentage columns.
Only categories present in at least 20\% of the data are included.
}
\description{
Convert categorical employee data into percentage-based columns by team and school year
}
