% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_enrollments_cat_val_to_pct_columns}
\alias{transform_enrollments_cat_val_to_pct_columns}
\title{Transform Categorical Values to Percentage Columns for Enrollments}
\usage{
transform_enrollments_cat_val_to_pct_columns(enrollments)
}
\arguments{
\item{enrollments}{A data frame containing enrollment data with TEAM_naam
and COHORT_naam columns.}
}
\value{
A data frame with categorical variables converted to percentage columns.
}
\description{
Convert categorical variables in enrollment data to percentage columns,
grouped by team and cohort
}
