% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_to_cat_val_pct_columns}
\alias{transform_to_cat_val_pct_columns}
\title{Transform Categorical Variables into Value-Percentage Columns}
\usage{
transform_to_cat_val_pct_columns(
  data,
  grouping_vars = c("TEAM_naam", "COHORT_naam"),
  max_n_values = 6
)
}
\arguments{
\item{data}{A data frame containing categorical variables.}

\item{grouping_vars}{Optional. A character vector of column names to group by. Defaults to \code{c("TEAM_naam", "COHORT_naam")}.}

\item{max_n_values}{Optional. Maximum number of distinct values to consider for processing. Defaults to 6.}
}
\value{
A data frame with categorical variables transformed into value count and percentage columns,
grouped by the specified variables.
}
\description{
Convert categorical variables into columns showing value counts and percentages
}
