% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{pivot_cat_values_to_pct}
\alias{pivot_cat_values_to_pct}
\title{Pivot Categorical Values to Percentages}
\usage{
pivot_cat_values_to_pct(data, col_name)
}
\arguments{
\item{data}{A data frame containing the categorical column to pivot}

\item{col_name}{A string specifying the column name to pivot}
}
\value{
A data frame with the categorical values pivoted to percentage columns.
Each unique value in \code{col_name} becomes a new column with the percentage
of occurrences within each group.
}
\description{
Transform categorical values into columns with percentages
}
