% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{select_best_params}
\alias{select_best_params}
\title{Select Best Parameters from Tuning Results}
\usage{
select_best_params(tuned_results)
}
\arguments{
\item{tuned_results}{Tuning results from tune_model()}
}
\value{
The best hyperparameter values
}
\description{
Select the best hyperparameters from tuning results.
}
